\name{biplot}
\alias{biplot}
\title{Generate a biplot of two variables}
\usage{
biplot(campaign_urn, prompt_id, prompt2_id, ...)
}
\arguments{
  \item{campaign_urn}{id of the campaign}

  \item{prompt_id}{prompt on the x axis}

  \item{prompt2_id}{prompt on the y axis}

  \item{...}{other parameters passed on to
  oh.survey_response/read}
}
\value{
ggplot2 plot object
}
\description{
Generate a biplot of two variables
}

