\name{distributionplot}
\alias{distributionplot}
\title{Shows a histogram or barchart of the data}
\usage{
  distributionplot(campaign_urn, prompt_id, ...)
}
\arguments{
  \item{campaign_urn}{campaign id}

  \item{prompt_id}{id of the prompt}

  \item{...}{other arguments passed to
  oh.survey_response.read}
}
\value{
  ggplot2 plot object
}
\description{
  Shows a histogram or barchart of the data
}

