\name{gmapdata}
\alias{gmapdata}
\title{Not a plot, but generates some data with geotags for the demo front-end}
\usage{
gmapdata(prompt_id = NULL, ...)
}
\arguments{
  \item{prompt_id}{an optional prompt id to be added to the
  output}

  \item{...}{arguments passed on to
  oh.survey_response.read}
}
\value{
a list with geotags and optionally pictures
}
\description{
Not a plot, but generates some data with geotags for the
demo front-end
}

