\name{sharedplot}
\alias{sharedplot}
\title{A barchart of the number of shared and unshared responses per campaign}
\usage{
sharedplot(campaign_urn, ...)
}
\arguments{
  \item{campaign_urn}{campaign id}

  \item{...}{arguments passed on to
  oh.survey_response.read}
}
\value{
a ggplot2 object
}
\description{
A barchart of the number of shared and unshared responses
per campaign
}

