\name{userplot}
\alias{userplot}
\title{Timeseries plot of data for a single user}
\usage{
  userplot(campaign_urn, prompt_id, user_id, ...)
}
\arguments{
  \item{campaign_urn}{campaign id}

  \item{prompt_id}{prompt id}

  \item{user_id}{user id}

  \item{...}{arguments passed on to
  oh.survey_response.read}
}
\value{
  a ggplot2 plot object
}
\description{
  Timeseries plot of data for a single user
}

